pyproject.toml
[project]
name = "Pynzor"
version = "1.0.8"
description = "Web pentesting CLI - scan ports, fuzz directories, hunt headers, and probe for vulns"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "yal212", email = "yal212yal@gmail.com"}
]
keywords = ["security", "pentesting", "web", "reconnaissance", "scanner"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "typer[all]>=0.12",
    "httpx>=0.27",
    "rich>=13",
    "dnspython>=2.6",
    "beautifulsoup4>=4.12",
    "lxml>=5",
    "pyyaml>=6",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "pytest-cov>=5",
]

[project.urls]
Homepage = "https://github.com/yal212/Pynzor"
Repository = "https://github.com/yal212/Pynzor"

[project.scripts]
Pynzor = "cli.main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["reports"]

[tool.setuptools.package-data]
"cli" = ["config.yaml", "wordlists/*"]
"output" = ["templates/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["Pynzor"]
omit = ["tests/*", "*/conftest.py"]

[tool.pyright]
venv = ".venv"
